Remove run_as_system_user setting
I think for now we should require users to exist on the
machine. If we set this to false, we will be spawning
users as root currently (since we require jupyterhub to be
running as root) & do not have a very clear set of instructions
on how to do that safely. Doing that entirely safely might
also not be possible right now, because of deficiencies in
systemd-run (can not seem to drop Capabilities yet, for example).

So instead of having a hard to secure setting in here, let's
just drop it for now. We can bring it back if we have users
wanting to use systemd spawner without the users existing on
the local system.