Use username, not userid in unit_name_template
Makes it much easier, and we do not have to worry about
invalid cahracters right now because we require
that the users exist on the system with the same name already.
For times when this is not true and a user map of some sort
is being used, we'll have to FIXME to deal with those anyway.