Remove get_state and load_state methods
The renaming of templates will probably not be a common occurance and if they
happen one can just stop all affected services, remove their units and
daemon-reload. JupyterHub does not need knowledge of the implementation detail
what the units are named.

Having the unit name resurrected from some database field only leads to hard to
debug problems when state becomes split between what's in the filesystem and
what's in the database, which is too much logic in the spawner. It should just
start and stop units.